Refactor import statements in test files to use sammy_options module
